+++ calculator.h
@@ -1,0 +1,4 @@
+#pragma once
+#define PI 3.14159265359
+#define TOLERANCE 1.e-6
+#define DIM 3
