+++ src/algebra.h
@@ -1,0 +1,8 @@
+#pragma once
+
+double algeb_sum(double val1, double val2);
+double algeb_subs(double val1, double val2);
+double algeb_divs(double val1, double val2);
+double algeb_mult(double val1, double val2);
+double algeb_power(double base, double exponent);
+double algeb_square(double x, double n);
