+++ src/trigonometrics.h
@@ -1,0 +1,2 @@
+#pragma once
+float sine(float x);
