+++ src/vectors.h
@@ -1,0 +1,8 @@
+#pragma once
+void PrintVector(float *v);
+void SumVectors(float *v1, float *v2, float *sum);
+void SusVectors(float *v1, float *v2, float *sus);
+void ProVectors(float *v1, float *v2, float *pro);
+void DivVectors(float *v1, float *v2, float *Div);
+void copiArray(float w[], float wans[]);
+
